DeepReader/Table/requirements.txt
numpy==1.18
pandas==0.25.3
pdf2image
matplotlib==3.1.0
opencv-python==4.4.0.46
tensorflow-cpu==1.15.0
tabulate
google-cloud-storage
boto3
